+++ PN022BLL/biz/patient/patient.h
@@ -22,7 +22,7 @@ public:
 	static BizPatient^ BizPatient::Instance();
 
 	property String^			systemId;			// Customer System ID
-	property unsigned short		patientNumber;		// Internal patient number
+	property unsigned int		patientNumber;		// Internal patient number
 
 	property String^			groupStudyId;		// User defined group/study
 	property String^			patientId;			// User defined patient ID
